LU-17987 ofd: always trust OST last_id
In normal precreate, the last_id on OST is always trustable because
it's maintained in memory, and it should be MDT that messes up the
last created FID in reconnect. Even if the gap is more than
OST_MAX_PRECREATE, reply the last_id to MDT to avoid leaving OST in
unavailable status.

Signed-off-by: Lai Siyao <lai.siyao@whamcloud.com>
Change-Id: I838aa0ab15e6adaaa340356e54ab597a43a9fdf4
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55574
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Mikhail Pershin <mpershin@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>