LU-11077 utils: --client option for set_param
Added new [--client|-C[FSNAME]] option for 'lctl set_param' which
writes the parameter to the local /etc/lustre/mount.client.params
config file. Upon each Lustre client mount those parameters will
be set on the local node. If FSNAME was provided, the parameters
will be saved in the mount-specific /etc/lustre/mount.FSNAME.params
config file, and will be set (and override) the more generic client
mount parameters on that node when that filesystem is mounted.
However only parameters containing FSNAME can be set to their
respective params config file to avoid generic parameters that are
only supposed to affect a single filesystem, actually affecting all
of them.

Can be used together with [--delete|-d] to remove the parameter from
the given log file. If [--delete|-d] is specified without -C or -P
it will enable -P by default. A warning message will be printed when
this happens so users are aware of what's going on.

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: Iec0b9bfb5e259154ed2439e6e505b826a888905f
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55858
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>