LU-17922 utils: added idmap range functionality
Added the ability to a declare a range when adding idmaps to a
nodemap. The syntax is:
 <clientid_start>-<clientid_end>:<fsid_start>[-<fsid_end>]

The uid_end value is optional. In practice this looks like:

nodemap_add_idmap --name test --idtype uid --idmap 500-510:10000

It is also now possible to delete idmap ranges with the
nodemap_del_idmap command as well with the same syntax.

Signed-off-by: Maximilian Dilger <mdilger@whamcloud.com>
Change-Id: If6a2b9ab11f7d435a6854055001e6102aac43115
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55502
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>