LU-17501 libcfs: allow CPT exclude list for cores
Allow a relative CPU core exclude list to be specified when
constructing the CPT distribution instead of having to specify all of
the CPU cores except the ones that should be avoided. The CPU core
numbers are the relative core numbers to each NUMA node, such that
for each node, the Mth to Nth cores are excluded. The exclude list
is set by specifying 'C' in /etc/modprobe.d/lustre.conf for the CPT
number, with either a comma-separated list or a range of CPU cores:

    options libcfs cpu_pattern="N C[0-3]"

or:

    options libcfs cpu_pattern="C[0,1,2,3]"

Both options are equivalent. This would exclude cores 0-3 on each NUMA
node from use by LNet and Lustre.

It isn't possible to specify both include and exclude lists at the
same time, but it doesn't really make sense to do this anyway.

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: Ief5f36e0bbc49865317cee199c91c9b4a350c418
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55544
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>