LU-17986 tests: Update sanity-lnet 111 for IPv6
Modify test_111 to use setup_router_test() so that it will use
IPv6 NIDs for the --gateway argument when that is what the test
environment has configured.

The test is also modified to check success/failure of the route add
command, and to verify that the correct number for routes were added.

Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I8dea5874dd01eb7f35603bed46660cd6c6b85041
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55561
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>