LU-17853 ptlrpc: Negative value for req_waittime
A negative value was being reported in req_waittime mdt. This was
likely caused by a backwards adjustment of a few microseconds and
would cause a negative time delta to be calculated.

Fixed negative wait times by setting the time delta to be 1 if
the time delta was calculated to be negative. This should have
minimal impact on the statistics.

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I48543b8b1fbc83829421a30f4f7be7da8b681132
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55605
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Nathaniel Clark <nclark@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>