LU-17431 nodemap: fileset inheritance for dynamic nodemap
Filesets for a dynamic nodemap are inherited from the parent nodemap.
It is possible to define a primary or alternate fileset specific to
the dynamic nodemap, but only if it is a subdirectory of the parent
nodemap's filesets, or if the parent has no filesets defined.
As we do not want to leave dynamic nodemaps without any primary
fileset defined, the only way currently available is to use
'lctl nodemap_set_fileset'. 'lctl nodemap_fileset_add/del' cannot be
used for the primary fileset of a dynamic nodemap.

Add sanity-sec test_72d to exercise this feature.

Test-Parameters: trivial
Test-Parameters: testgroup=review-dne-part-2
Test-Parameters: testgroup=review-dne-selinux-ssk-part-2
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I61a5fe82d92233bc66d10b8c1b939c04777bcf94
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55614
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>