LU-17928 lnet: add kmod devel package
This creates a new kernel module development package for building
kernel modules that depend on the Lustre/LNet kAPI

The most notable of these is DVS which uses LNet

Along with the kernel includes add a package config file: lnet.pc
and the Module.symvers needed for linking against Lustre/LNet kAPI

Use:
   pkg-config --variable=symversdir lnet
to find the path to Module.symvers and include files.

In addition the dkms build can differ enough that the packaged
Module.symvers and config.h (and possibly the headers as well)
may diff enough that they are not interchangeable.

Use the update-alternatives subsystem to enable the dkms and kmp
packages to co-exist and the kmp-devel package to work with either.

Also loosens user space requirement to require:
 Lustre version >= major.minor
and not the exact build

Test-Parameters: trivial
HPE-bug-id: LUS-12246, LUS-12378, LUS-12351
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Signed-off-by: Caleb Carlson <caleb.carlson@hpe.com>
Change-Id: Idb00b881e8f6d4a703cc71fd0d8768e1f433fca3
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55387
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>