LU-14442 utils: --tunable opt for list/get_param
Added new [--tunable|-t] option for 'lctl {list|get}_param' that
will only print tunable parameters. Specifically it excludes all
params that contain any of the following strings:

console|debug_|fail_|force|import|nis|panic_|sepol|stats|target_obd

This is used for debugging system performance as it creates an easy
way to collect all of the Lustre tuning parameters for each run.
This can be combined with --no-links and --readable to remove a great
deal of unwanted information being printed.

LIST_PATHNAME was removed as it only added more to maintain while
the purpose of that option can be achieved in a single line.

Test-Parameters:trivial
Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I847884af2fc16611f34e3e44a3f3990d39d7f39b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55723
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Olaf Faaland <faaland1@llnl.gov>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>