LU-18032 llite: disable unaligned AIO
Parallel copy for unaligned DIO has an unpleasant
interaction with AIO - Since AIO is async, we must do the
copying first.

This shows up as rare failures in 119i, which tests
unaligned AIO.

Handling this correctly for AIO introduces some complexity,
so instead we disable unaligned AIO.  This is a very small
use case and the complexity is not worth the benefit.

Signed-off-by: Patrick Farrell <patrick.farrell@oracle.com>
Change-Id: I81781da841b83130c9a7b5855abd88ae048f1908
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55716
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>