LU-18012 obd: support large NIDs when MGS is down
When the mgc is initialized it will create an export for the MGS.
Having a MGS export does not mean we can actually reach the
remote MGS. The way to know if the MGS is reachable is to examine
the connect flags after obd_connect() is called. If its zero
then the remote MGS is down. If this is the case since when
starting the mgc we examine the mgsnode= string we can examine
every NID. If one of those NIDs for mgsnode= or the failover
has a large NID address then force support for large NIDs.

Change-Id: I3550df21618f07b69701a9d337d3dfc5f99e49a3
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55907
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>